refactor(domain): decompose interactive-session-service into 5 focused classes
Extract 1200+ line InteractiveSessionService into focused sub-classes:
- SessionState: shared types and constants
- SubscriberNotifier: real-time subscription management
- SessionStateManager: session CRUD, stop, markRead, turn statuses
- ChatStateBuilder: chat UI state reconstruction from DB + memory
- TurnExecutor: turn execution, stream processing, user interactions
- SessionBootSequence: session init, cap enforcement, agent boot

The original class is now a thin facade delegating to sub-classes,
preserving all existing call sites and the IInteractiveSessionService
interface contract. All 41 unit tests pass unchanged.
