+++ packages/core/src/infrastructure/services/interactive/interactive-session.service.ts
@@ -1,87 +1,13 @@
 /**
- * Interactive Session Service
+ * Interactive Session Service (Facade)
  *
- * Singleton service that owns the lifecycle of all interactive agent sessions.
- * Uses the IAgentExecutorFactory to create interactive executors that manage
- * persistent sessions via the agent SDK. Multi-turn context is maintained
- * by the SDK session handle internally.
+ * Thin facade that delegates to focused sub-classes:
+ * - SubscriberNotifier — real-time subscription management
+ * - SessionStateManager — session CRUD, stop, markRead, turn statuses
+ * - ChatStateBuilder — chat UI state reconstruction
+ * - TurnExecutor — turn execution, message persistence, user interactions
+ * - SessionBootSequence — session initialization and agent boot
  *
- * Dependencies are injected via constructor for testability (no real processes
- * are spawned in unit tests — the factory is replaced with a test double).
- */
-
-import * as crypto from 'node:crypto';
-import type {
-  IInteractiveSessionService,
-  StreamChunk,
-  UnsubscribeFn,
-  ChatState,
-} from '../../../application/ports/output/services/interactive-session-service.interface.js';
-import type { IInteractiveSessionRepository } from '../../../application/ports/output/repositories/interactive-session-repository.interface.js';
-import type { IInteractiveMessageRepository } from '../../../application/ports/output/repositories/interactive-message-repository.interface.js';
-import type { IAgentExecutorFactory } from '../../../application/ports/output/agents/agent-executor-factory.interface.js';
-import type {
-  InteractiveAgentSessionHandle,
-  UserInteractionData,
-} from '../../../application/ports/output/agents/interactive-agent-executor.interface.js';
-import type { IFeatureRepository } from '../../../application/ports/output/repositories/feature-repository.interface.js';
-import type { InteractiveSession, InteractiveMessage } from '../../../domain/generated/output.js';
-import {
-  InteractiveSessionStatus,
-  InteractiveMessageRole,
-  AgentType,
-  AgentAuthMethod,
-} from '../../../domain/generated/output.js';
-import type { AgentConfig } from '../../../domain/generated/output.js';
-import { ConcurrentSessionLimitError } from '../../../domain/errors/concurrent-session-limit.error.js';
-import { type FeatureContextBuilder } from './feature-context.builder.js';
-import { getSettings, hasSettings } from '../settings.service.js';
-
-/** Default idle timeout if no settings are loaded (15 minutes). */
-const DEFAULT_TIMEOUT_MS = 15 * 60 * 1000;
-
-/** Default concurrent session cap. */
-const DEFAULT_CAP = 3;
-
-/** Maximum time to wait for the agent to become ready (60 seconds). */
-const BOOT_TIMEOUT_MS = 60_000;
-
-/** In-memory state for a single live session. */
-interface SessionState {
-  sessionId: string;
-  featureId: string;
-  worktreePath: string;
-  /** Agent SDK session handle — null until session is created. */
-  handle: InteractiveAgentSessionHandle | null;
-  /** Agent SDK session ID for resumption across service restarts. */
-  agentSessionId?: string;
-  timer: NodeJS.Timeout | null;
-  /** Accumulates assistant text between user turns for persistence. */
-  currentAssistantBuffer: string;
-  /** Accumulates tool events during a turn for rich message persistence. */
-  toolEventsLog: string[];
-  /** Subscriber callbacks for real-time stdout chunk forwarding. */
-  subscribers: Set<(chunk: StreamChunk) => void>;
-  /** User message content queued while session boots. */
-  pendingUserContent?: string;
-  /** Model override for the agent process (e.g. 'claude-sonnet-4-6'). */
-  model?: string;
-  /** Agent type for this session. */
-  agentType?: string;
-  /** AbortController to cancel active stream iteration on stop. */
-  streamAbort?: AbortController;
-  /** Whether a turn is currently executing (prevents concurrent turns). */
-  turnInProgress: boolean;
-  /** Queue of user messages waiting to be sent after the current turn completes. */
-  turnQueue: string[];
-  /** Pending user interaction (AskUserQuestion) — agent stream is paused, waiting for response. */
-  pendingInteraction: UserInteractionData | null;
-  /** Resolver for the pending interaction Promise — call to resume the agent. */
-  pendingInteractionResolver: ((answers: Record<string, string>) => void) | null;
-}
-
-/**
- * Core service managing interactive agent session lifecycles.
  * Must be registered as a singleton in the DI container.
  *
  * **Polymorphic `featureId` scope key:** The `featureId` parameter accepted
@@ -92,33 +18,71 @@ interface SessionState {
  * - Global chat: literal string `"global"`
  *
  * Sessions and messages are isolated by this key regardless of chat type.
- *
- * @todo Consider renaming to `scopeId` + adding a `scopeType` discriminator.
  */
+
+import type {
+  IInteractiveSessionService,
+  StreamChunk,
+  UnsubscribeFn,
+  ChatState,
+} from '../../../application/ports/output/services/interactive-session-service.interface.js';
+import type { IInteractiveSessionRepository } from '../../../application/ports/output/repositories/interactive-session-repository.interface.js';
+import type { IInteractiveMessageRepository } from '../../../application/ports/output/repositories/interactive-message-repository.interface.js';
+import type { IAgentExecutorFactory } from '../../../application/ports/output/agents/agent-executor-factory.interface.js';
+import type { IFeatureRepository } from '../../../application/ports/output/repositories/feature-repository.interface.js';
+import type { InteractiveSession, InteractiveMessage } from '../../../domain/generated/output.js';
+import type { FeatureContextBuilder } from './feature-context.builder.js';
+import type { SessionState } from './session-state.js';
+import { DEFAULT_TIMEOUT_MS } from './session-state.js';
+import { SubscriberNotifier } from './subscriber-notifier.js';
+import { SessionStateManager } from './session-state-manager.js';
+import { ChatStateBuilder } from './chat-state-builder.js';
+import { TurnExecutor } from './turn-executor.js';
+import { SessionBootSequence } from './session-boot-sequence.js';
+import { getSettings, hasSettings } from '../settings.service.js';
+
 export class InteractiveSessionService implements IInteractiveSessionService {
   /** Live sessions indexed by sessionId. */
   private sessions = new Map<string, SessionState>();
   /** Cached agentSessionIds from stopped sessions, keyed by featureId. */
   private stoppedAgentSessionIds = new Map<string, string>();
-  /**
-   * Feature-level subscribers that survive session restarts.
-   *
-   * Unlike session-level subscribers (in SessionState.subscribers), these
-   * persist when a session dies and a new one boots. SSE connections
-   * subscribe here so they continue receiving events from new sessions.
-   */
-  private featureSubscribers = new Map<string, Set<(chunk: StreamChunk) => void>>();
+
+  // Sub-services
+  private readonly notifier: SubscriberNotifier;
+  private readonly stateManager: SessionStateManager;
+  private readonly chatStateBuilder: ChatStateBuilder;
+  private readonly turnExecutor: TurnExecutor;
+  private readonly bootSequence: SessionBootSequence;
 
   constructor(
-    private readonly sessionRepo: IInteractiveSessionRepository,
-    private readonly messageRepo: IInteractiveMessageRepository,
-    private readonly executorFactory: IAgentExecutorFactory,
-    private readonly featureRepo: IFeatureRepository,
-    private readonly contextBuilder: FeatureContextBuilder
-  ) {}
+    sessionRepo: IInteractiveSessionRepository,
+    messageRepo: IInteractiveMessageRepository,
+    executorFactory: IAgentExecutorFactory,
+    featureRepo: IFeatureRepository,
+    contextBuilder: FeatureContextBuilder
+  ) {
+    this.notifier = new SubscriberNotifier();
+    this.stateManager = new SessionStateManager(sessionRepo, messageRepo);
+    this.chatStateBuilder = new ChatStateBuilder(sessionRepo, messageRepo);
+
+    // Bind resetTimer so sub-classes can use it
+    const resetTimer = (state: SessionState): void => this.resetTimer(state);
+
+    this.turnExecutor = new TurnExecutor(sessionRepo, messageRepo, this.notifier, resetTimer);
+    this.bootSequence = new SessionBootSequence(
+      sessionRepo,
+      messageRepo,
+      executorFactory,
+      featureRepo,
+      contextBuilder,
+      this.notifier,
+      this.turnExecutor,
+      resetTimer
+    );
+  }
 
   // ---------------------------------------------------------------------------
-  // Public API
+  // Public API — delegates to sub-classes
   // ---------------------------------------------------------------------------
 
   async startSession(
@@ -127,725 +91,48 @@ export class InteractiveSessionService implements IInteractiveSessionService {
     model?: string,
     agentType?: string
   ): Promise<InteractiveSession> {
-    const cap = this.getCap();
-    const activeCount = await this.sessionRepo.countActiveSessions();
-    if (activeCount >= cap) {
-      throw new ConcurrentSessionLimitError(activeCount, cap);
-    }
-
-    // Create DB record with booting status
-    const now = new Date();
-    const session: InteractiveSession = {
-      id: crypto.randomUUID(),
-      featureId,
-      status: InteractiveSessionStatus.booting,
-      startedAt: now,
-      lastActivityAt: now,
-      createdAt: now,
-      updatedAt: now,
-    };
-    await this.sessionRepo.create(session);
-
-    // Mark as processing immediately so the FAB shows the spinner during boot
-    void this.sessionRepo.updateTurnStatus(session.id, 'processing');
-
-    // Carry over agentSessionId from previous session so resumption works
-    let previousAgentSessionId: string | undefined;
-    for (const [, s] of this.sessions) {
-      if (s.featureId === featureId && s.agentSessionId) {
-        previousAgentSessionId = s.agentSessionId;
-        break;
-      }
-    }
-    // Also check stoppedSessions cache (populated on stop)
-    previousAgentSessionId ??= this.stoppedAgentSessionIds.get(featureId);
-    // Fall back to DB — the in-memory cache may be empty after service restart
-    if (!previousAgentSessionId) {
-      const latestDbSession = await this.sessionRepo.findByFeatureId(featureId);
-      if (latestDbSession) {
-        previousAgentSessionId =
-          (await this.sessionRepo.getAgentSessionId(latestDbSession.id)) ?? undefined;
-      }
-    }
-
-    // Set up in-memory state
-    const state: SessionState = {
-      sessionId: session.id,
+    return this.bootSequence.startSession(
       featureId,
       worktreePath,
+      this.sessions,
+      this.stoppedAgentSessionIds,
       model,
-      agentType,
-      handle: null,
-      agentSessionId: previousAgentSessionId,
-      timer: null,
-      currentAssistantBuffer: '',
-      toolEventsLog: [],
-      subscribers: new Set(),
-      turnInProgress: false,
-      turnQueue: [],
-      pendingInteraction: null,
-      pendingInteractionResolver: null,
-    };
-    this.sessions.set(session.id, state);
-
-    // Fire-and-forget the async boot sequence. The API returns the session
-    // immediately in "booting" status; the frontend polls until "ready".
-    void this.completeBootAsync(state, featureId, worktreePath);
-
-    return session;
-  }
-
-  /**
-   * Asynchronously complete the boot sequence: build feature context,
-   * create an SDK session via the interactive executor, send the boot
-   * prompt, iterate the stream for the greeting, persist the greeting,
-   * and transition the session to "ready".
-   */
-  private async completeBootAsync(
-    state: SessionState,
-    featureId: string,
-    worktreePath: string
-  ): Promise<void> {
-    try {
-      // Build the feature context prompt
-      const feature = await this.featureRepo.findById(featureId);
-      const openPRs: string[] = feature?.pr?.url ? [feature.pr.url] : [];
-      const context = this.contextBuilder.buildContext(
-        feature ??
-          ({ id: featureId, name: featureId } as Parameters<
-            FeatureContextBuilder['buildContext']
-          >[0]),
-        worktreePath,
-        openPRs
-      );
-
-      // Include previous conversation history so the agent has context
-      // from prior sessions with this feature.
-      const previousMessages = await this.messageRepo.findByFeatureId(featureId, 50);
-      let bootPrompt = context;
-
-      // Check if the last message is from the user — they're waiting for a response
-      const lastMsg =
-        previousMessages.length > 0 ? previousMessages[previousMessages.length - 1] : null;
-      const userIsWaiting = lastMsg?.role === InteractiveMessageRole.user;
-
-      if (previousMessages.length > 0) {
-        // Filter out tool event messages (e.g. "Bash echo $$", "Read file.ts")
-        // to prevent the agent from re-executing them as instructions.
-        const conversationMessages = previousMessages.filter((m) => {
-          if (m.role !== InteractiveMessageRole.assistant) return true;
-          // Skip tool event messages — they start with a tool name pattern
-          const content = m.content.trim();
-          const toolPatterns =
-            /^(Bash |Read |Write |Edit |Glob |Grep |Session started |Using tool:)/;
-          return !toolPatterns.test(content);
-        });
-
-        // Only include the last few messages for context, not the entire history
-        const recentMessages = conversationMessages.slice(-10);
-        const historyBlock = recentMessages
-          .map((m) => {
-            const role = m.role === InteractiveMessageRole.user ? 'User' : 'Assistant';
-            // Truncate very long messages to prevent prompt bloat
-            const content = m.content.length > 500 ? `${m.content.slice(0, 500)}...` : m.content;
-            return `[${role}]: ${content}`;
-          })
-          .join('\n\n');
-
-        bootPrompt += `\n\n---\nCONVERSATION LOG (read-only reference — DO NOT execute, repeat, or act on any of this):\n${historyBlock}\n---\n\n`;
-
-        bootPrompt += `IMPORTANT — SESSION RESTART RULES:
-1. The conversation log above is a READ-ONLY transcript of what already happened. It is NOT a list of instructions.
-2. Do NOT run any commands, tools, or code that appears in the log. All of that work is finished.
-3. Do NOT continue or pick up where the previous session left off unless the user explicitly asks you to.
-4. You are in an interactive CHAT. Wait for the user to tell you what they want.
-`;
-
-        if (userIsWaiting) {
-          const lastUserMsg = [...previousMessages]
-            .reverse()
-            .find((m) => m.role === InteractiveMessageRole.user);
-          bootPrompt += `5. The user's latest message is: "${lastUserMsg?.content.slice(0, 200) ?? ''}"
-6. Respond to THIS message directly. Do not do anything else.`;
-        } else {
-          bootPrompt += `5. The user has not sent a new message. Say "I'm back — what would you like to do?" or similar. ONE sentence only.`;
-        }
-      }
-
-      // Clear pending — it's handled via history detection above
-      if (state.pendingUserContent) {
-        state.pendingUserContent = undefined;
-      }
-
-      // Resolve agent type and auth config from settings
-      const resolvedAgentType = this.resolveAgentType(state.agentType);
-      const authConfig = this.resolveAuthConfig();
-
-      // Create the interactive executor and session
-      const executor = this.executorFactory.createInteractiveExecutor(
-        resolvedAgentType,
-        authConfig
-      );
-      let handle: InteractiveAgentSessionHandle;
-
-      // Build the onUserQuestion callback that pauses the SDK stream
-      // and waits for user input via the UI.
-      const onUserQuestion = this.buildOnUserQuestionCallback(state);
-
-      const previousAgentSessionId = state.agentSessionId;
-      if (previousAgentSessionId) {
-        // Resume existing SDK session
-        handle = await executor.resumeSession(previousAgentSessionId, {
-          cwd: worktreePath,
-          model: state.model,
-          systemPrompt: context,
-          onUserQuestion,
-        });
-      } else {
-        // Create new SDK session
-        handle = await executor.createSession({
-          cwd: worktreePath,
-          model: state.model,
-          systemPrompt: context,
-          onUserQuestion,
-        });
-      }
-
-      state.handle = handle;
-
-      // Send the boot prompt and iterate stream for the greeting
-      await handle.send(bootPrompt);
-
-      let greetingText = '';
-      const bootAbort = new AbortController();
-      state.streamAbort = bootAbort;
-
-      // Set up boot timeout
-      const bootTimeout = setTimeout(() => {
-        bootAbort.abort();
-      }, BOOT_TIMEOUT_MS);
-
-      try {
-        for await (const event of handle.stream()) {
-          if (bootAbort.signal.aborted) {
-            throw new Error(`Agent boot timed out after ${BOOT_TIMEOUT_MS / 1000}s`);
-          }
-
-          this.resetTimer(state);
-
-          switch (event.type) {
-            case 'delta':
-              if (event.content) {
-                greetingText += event.content;
-                state.currentAssistantBuffer += event.content;
-                this.notify(state, { delta: event.content!, done: false });
-              }
-              break;
-
-            case 'tool_use':
-              if (event.label) {
-                const toolLabel = event.label;
-                const toolDetail = event.detail;
-                void this.persistToolEvent(state, toolLabel, toolDetail);
-                this.notify(state, {
-                  delta: '',
-                  done: false,
-                  log: `Using tool: ${toolLabel}`,
-                  activity: { kind: 'tool_use', label: toolLabel, detail: toolDetail },
-                });
-              }
-              break;
-
-            case 'tool_result':
-              if (event.label) {
-                const resultLabel = event.label;
-                const resultDetail = event.detail;
-                void this.persistToolEvent(state, resultLabel, resultDetail);
-                this.notify(state, {
-                  delta: '',
-                  done: false,
-                  log: `Completed: ${resultLabel}`,
-                  activity: { kind: 'tool_result', label: resultLabel, detail: resultDetail },
-                });
-              }
-              break;
-
-            case 'status':
-              if (event.content) {
-                const statusContent = event.content;
-                this.notify(state, { delta: '', done: false, log: statusContent });
-              }
-              break;
-
-            case 'done': {
-              // Use result text if provided and non-empty, otherwise use accumulated buffer
-              const resultText =
-                event.content && event.content.length > 0 ? event.content : greetingText;
-
-              // Capture the SDK session ID (available after first message exchange)
-              const sdkSessionId = handle.sessionId;
-              if (sdkSessionId) {
-                // Detect CWD mismatch: if we tried to resume but got a different
-                // session ID, the SDK silently created a fresh session (typically
-                // because the cwd changed or session JSONL was lost).
-                if (previousAgentSessionId && sdkSessionId !== previousAgentSessionId) {
-                  // eslint-disable-next-line no-console
-                  console.warn(
-                    `[InteractiveSession] Session resume mismatch for feature ${featureId}: ` +
-                      `expected ${previousAgentSessionId}, got ${sdkSessionId}. ` +
-                      `SDK created a fresh session (likely cwd changed or session expired).`
-                  );
-                }
-                state.agentSessionId = sdkSessionId;
-                // Persist to DB so it survives service restarts
-                void this.sessionRepo.updateAgentSessionId(state.sessionId, sdkSessionId);
-              }
-
-              // Persist greeting and mark session ready
-              const greetingMsg: InteractiveMessage = {
-                id: crypto.randomUUID(),
-                featureId,
-                sessionId: state.sessionId,
-                role: InteractiveMessageRole.assistant,
-                content: resultText,
-                createdAt: new Date(),
-                updatedAt: new Date(),
-              };
-              await this.messageRepo.create(greetingMsg);
-              await this.sessionRepo.updateStatus(state.sessionId, InteractiveSessionStatus.ready);
-
-              // If there's a pending user message, the next turn will set 'processing'.
-              // Otherwise boot greeting is expected — mark idle.
-              if (!state.pendingUserContent) {
-                void this.sessionRepo.updateTurnStatus(state.sessionId, 'idle');
-              }
-
-              state.currentAssistantBuffer = '';
-              state.toolEventsLog = [];
-
-              // Notify subscribers of end-of-turn
-              this.notify(state, { delta: '', done: true });
-
-              // Start idle timer now that the session is live
-              this.resetTimer(state);
-              return; // Boot complete
-            }
-
-            case 'error':
-              throw new Error(`Agent error during boot: ${event.content ?? 'unknown'}`);
-          }
-        }
-      } finally {
-        clearTimeout(bootTimeout);
-        state.streamAbort = undefined;
-      }
-
-      // If we get here without a 'done' event, use whatever text we accumulated
-      if (greetingText) {
-        const greetingMsg: InteractiveMessage = {
-          id: crypto.randomUUID(),
-          featureId,
-          sessionId: state.sessionId,
-          role: InteractiveMessageRole.assistant,
-          content: greetingText,
-          createdAt: new Date(),
-          updatedAt: new Date(),
-        };
-        await this.messageRepo.create(greetingMsg);
-      }
-      await this.sessionRepo.updateStatus(state.sessionId, InteractiveSessionStatus.ready);
-      if (!state.pendingUserContent) {
-        void this.sessionRepo.updateTurnStatus(state.sessionId, 'idle');
-      }
-      state.currentAssistantBuffer = '';
-      state.toolEventsLog = [];
-      this.resetTimer(state);
-    } catch (err) {
-      // If session was already cleaned up by stopSession, nothing more to do
-      if (!this.sessions.has(state.sessionId)) return;
-
-      // Boot failed — mark session as error so the frontend can show the failure
-      // eslint-disable-next-line no-console
-      console.error(`[InteractiveSession] boot failed for session ${state.sessionId}:`, err);
-      try {
-        await this.sessionRepo.updateStatus(state.sessionId, InteractiveSessionStatus.error);
-      } catch {
-        // Best-effort DB update
-      }
-      if (state.agentSessionId) {
-        this.stoppedAgentSessionIds.set(state.featureId, state.agentSessionId);
-      }
-      this.sessions.delete(state.sessionId);
-    }
+      agentType
+    );
   }
 
   async stopSession(sessionId: string): Promise<void> {
-    const state = this.sessions.get(sessionId);
-    if (!state) {
-      // Already stopped — idempotent
-      return;
-    }
-
-    // eslint-disable-next-line no-console
-    console.log(
-      `[InteractiveSession] stopSession called for ${sessionId} (feature: ${state.featureId})`,
-      new Error().stack?.split('\n').slice(1, 4).join(' <- ')
-    );
-
-    // Abort any active stream iteration and clear pending turns
-    if (state.streamAbort) {
-      state.streamAbort.abort();
-      state.streamAbort = undefined;
-    }
-    state.turnQueue.length = 0;
-    state.turnInProgress = false;
-
-    this.clearTimer(state);
-    // Cache agentSessionId so resumption works when session restarts
-    if (state.agentSessionId) {
-      this.stoppedAgentSessionIds.set(state.featureId, state.agentSessionId);
-    }
-    this.sessions.delete(sessionId);
-
-    // Close the SDK session handle
-    if (state.handle) {
-      try {
-        await state.handle.close();
-      } catch {
-        // Session may already be closed
-      }
-      state.handle = null;
-    }
-
-    await this.sessionRepo.updateStatus(sessionId, InteractiveSessionStatus.stopped, new Date());
-    void this.sessionRepo.updateTurnStatus(sessionId, 'idle');
+    return this.stateManager.stopSession(sessionId, this.sessions, this.stoppedAgentSessionIds);
   }
 
   async sendMessage(sessionId: string, content: string): Promise<InteractiveMessage> {
-    const dbSession = await this.sessionRepo.findById(sessionId);
-    if (!dbSession || dbSession.status !== InteractiveSessionStatus.ready) {
-      throw new Error(`Session ${sessionId} is not ready — cannot send message`);
-    }
-
-    const state = this.sessions.get(sessionId);
-    if (!state) {
-      throw new Error(`Session ${sessionId} is not ready — cannot send message`);
-    }
-
-    // Persist user message
-    const now = new Date();
-    const message: InteractiveMessage = {
-      id: crypto.randomUUID(),
-      featureId: state.featureId,
+    return this.turnExecutor.sendMessage(
       sessionId,
-      role: InteractiveMessageRole.user,
       content,
-      createdAt: now,
-      updatedAt: now,
-    };
-    await this.messageRepo.create(message);
-
-    // Reset idle timer on user activity
-    this.resetTimer(state);
-    await this.sessionRepo.updateLastActivity(sessionId, now);
-
-    // Guard: only one turn at a time per session (SDK stream is not concurrent-safe)
-    if (state.turnInProgress) {
-      state.turnQueue.push(content);
-    } else {
-      state.turnInProgress = true;
-      void this.executeAndPersistTurn(state, content);
-    }
-
-    return message;
-  }
-
-  /**
-   * Execute a turn via the SDK session handle and persist the assistant response.
-   */
-  private async executeAndPersistTurn(state: SessionState, prompt: string): Promise<void> {
-    try {
-      if (!state.handle) {
-        throw new Error('No active session handle — cannot execute turn');
-      }
-
-      state.currentAssistantBuffer = '';
-      state.toolEventsLog = [];
-
-      // Mark turn as processing for dot indicator
-      void this.sessionRepo.updateTurnStatus(state.sessionId, 'processing');
-
-      // Send the message to the SDK session
-      await state.handle.send(prompt);
-
-      // Set up abort controller for this stream
-      const abort = new AbortController();
-      state.streamAbort = abort;
-
-      let responseText = '';
-
-      try {
-        for await (const event of state.handle.stream()) {
-          if (abort.signal.aborted) break;
-
-          // Reset idle timer on each event received
-          this.resetTimer(state);
-
-          switch (event.type) {
-            case 'delta':
-              if (event.content) {
-                responseText += event.content;
-                state.currentAssistantBuffer += event.content;
-                this.notify(state, { delta: event.content!, done: false });
-              }
-              break;
-
-            case 'tool_use':
-              if (event.label) {
-                const toolLabel = event.label;
-                const toolDetail = event.detail;
-                void this.persistToolEvent(state, toolLabel, toolDetail);
-                this.notify(state, {
-                  delta: '',
-                  done: false,
-                  log: `Using tool: ${toolLabel}`,
-                  activity: { kind: 'tool_use', label: toolLabel, detail: toolDetail },
-                });
-              }
-              break;
-
-            case 'tool_result':
-              if (event.label) {
-                const resultLabel = event.label;
-                const resultDetail = event.detail;
-                void this.persistToolEvent(state, resultLabel, resultDetail);
-                this.notify(state, {
-                  delta: '',
-                  done: false,
-                  log: `Completed: ${resultLabel}`,
-                  activity: { kind: 'tool_result', label: resultLabel, detail: resultDetail },
-                });
-              }
-              break;
-
-            case 'status':
-              if (event.content) {
-                const statusContent = event.content;
-                this.notify(state, { delta: '', done: false, log: statusContent });
-              }
-              break;
-
-            case 'done': {
-              // Use result text if provided and non-empty, otherwise use accumulated buffer
-              const resultText =
-                event.content && event.content.length > 0 ? event.content : responseText;
-
-              // Persist assistant message
-              const now = new Date();
-              const msg: InteractiveMessage = {
-                id: crypto.randomUUID(),
-                featureId: state.featureId,
-                sessionId: state.sessionId,
-                role: InteractiveMessageRole.assistant,
-                content: resultText,
-                createdAt: now,
-                updatedAt: now,
-              };
-              await this.messageRepo.create(msg);
-
-              state.currentAssistantBuffer = '';
-              state.toolEventsLog = [];
-
-              // Accumulate usage from this turn
-              if (event.usage) {
-                void this.sessionRepo.accumulateUsage(state.sessionId, {
-                  costUsd: event.usage.costUsd ?? 0,
-                  inputTokens: event.usage.inputTokens ?? 0,
-                  outputTokens: event.usage.outputTokens ?? 0,
-                  turns: event.usage.numTurns ?? 1,
-                });
-              }
-
-              // Mark as unread — if user has the chat open, the frontend
-              // will immediately call markRead to clear it
-              void this.sessionRepo.updateTurnStatus(state.sessionId, 'unread');
-
-              // Notify subscribers of end-of-turn
-              this.notify(state, { delta: '', done: true });
-              return; // Turn complete
-            }
-
-            case 'error':
-              // eslint-disable-next-line no-console
-              console.error(
-                `[InteractiveSession] agent error during turn for session ${state.sessionId}:`,
-                event.content
-              );
-              // Accumulate usage even on errors — cost was still incurred
-              if (event.usage) {
-                void this.sessionRepo.accumulateUsage(state.sessionId, {
-                  costUsd: event.usage.costUsd ?? 0,
-                  inputTokens: event.usage.inputTokens ?? 0,
-                  outputTokens: event.usage.outputTokens ?? 0,
-                  turns: event.usage.numTurns ?? 1,
-                });
-              }
-              this.notify(state, {
-                delta: '',
-                done: true,
-                log: `Error: ${event.content ?? 'unknown'}`,
-              });
-              break;
-
-            case 'init':
-              // The SDK emits init on every turn, but we only show "Session started"
-              // during boot (handled in completeBootAsync). Ignore it here to avoid
-              // spamming the chat with repeated session-started messages.
-              break;
-
-            case 'api_retry':
-              this.notify(state, {
-                delta: '',
-                done: false,
-                log: event.content ?? 'Retrying API call...',
-              });
-              break;
-
-            case 'rate_limit':
-              this.notify(state, { delta: '', done: false, log: event.content ?? 'Rate limited' });
-              break;
-
-            case 'task_started':
-              if (event.content) {
-                void this.persistToolEvent(state, 'Subtask started', event.content);
-                this.notify(state, {
-                  delta: '',
-                  done: false,
-                  log: `Subtask: ${event.content}`,
-                  activity: { kind: 'system', label: 'Subtask started', detail: event.content },
-                });
-              }
-              break;
-
-            case 'task_progress':
-              if (event.content) {
-                this.notify(state, { delta: '', done: false, log: `Subtask: ${event.content}` });
-              }
-              break;
-
-            case 'task_done':
-              if (event.content) {
-                const taskStatus = event.detail ?? 'completed';
-                void this.persistToolEvent(state, `Subtask ${taskStatus}`, event.content);
-                this.notify(state, {
-                  delta: '',
-                  done: false,
-                  log: `Subtask ${taskStatus}: ${event.content}`,
-                  activity: {
-                    kind: 'system',
-                    label: `Subtask ${taskStatus}`,
-                    detail: event.content,
-                  },
-                });
-              }
-              break;
-
-            case 'user_question':
-              // AskUserQuestion is now handled by the canUseTool callback
-              // (buildOnUserQuestionCallback) which pauses the SDK stream.
-              // This event should not appear in the stream anymore, but if it
-              // does (e.g. from a different code path), ignore it here.
-              break;
-          }
-        }
-      } finally {
-        state.streamAbort = undefined;
-      }
-
-      // If we exit the stream loop without a 'done' event (stream ended),
-      // persist whatever text we accumulated
-      if (responseText && state.currentAssistantBuffer) {
-        const now = new Date();
-        const msg: InteractiveMessage = {
-          id: crypto.randomUUID(),
-          featureId: state.featureId,
-          sessionId: state.sessionId,
-          role: InteractiveMessageRole.assistant,
-          content: responseText,
-          createdAt: now,
-          updatedAt: now,
-        };
-        await this.messageRepo.create(msg);
-
-        state.currentAssistantBuffer = '';
-        state.toolEventsLog = [];
-        this.notify(state, { delta: '', done: true });
-      } else if (!responseText) {
-        // Stream ended without any response — SDK session likely died.
-        // Mark as error so the next message triggers a fresh session.
-        // eslint-disable-next-line no-console
-        console.error(
-          `[InteractiveSession] stream ended without response for session ${state.sessionId} — session may have died`
-        );
-        this.notify(state, {
-          delta: '',
-          done: true,
-          log: 'Session disconnected — will restart on next message',
-        });
-        if (state.agentSessionId) {
-          this.stoppedAgentSessionIds.set(state.featureId, state.agentSessionId);
-        }
-        this.sessions.delete(state.sessionId);
-        try {
-          await this.sessionRepo.updateStatus(state.sessionId, InteractiveSessionStatus.error);
-        } catch {
-          // Best-effort DB update
-        }
-        return; // Skip queue drain — session is dead
-      }
-    } catch (err) {
-      // If session was already stopped, ignore
-      if (!this.sessions.has(state.sessionId)) return;
-      // eslint-disable-next-line no-console
-      console.error(`[InteractiveSession] turn failed for session ${state.sessionId}:`, err);
-    } finally {
-      // Release the turn lock and drain the queue
-      state.turnInProgress = false;
-      if (this.sessions.has(state.sessionId) && state.turnQueue.length > 0) {
-        const nextContent = state.turnQueue.shift()!;
-        state.turnInProgress = true;
-        void this.executeAndPersistTurn(state, nextContent);
-      }
-    }
+      this.sessions,
+      this.stoppedAgentSessionIds
+    );
   }
 
   async getMessages(featureId: string, limit?: number): Promise<InteractiveMessage[]> {
-    return this.messageRepo.findByFeatureId(featureId, limit);
+    return this.stateManager.getMessages(featureId, limit);
   }
 
   async clearMessages(featureId: string): Promise<void> {
-    // Stop any active session so the agent doesn't retain old context
-    const state = this.findActiveStateForFeature(featureId);
-    if (state) {
-      await this.stopSession(state.sessionId);
-    }
-    // Also clear the cached agentSessionId so next session starts fresh
-    this.stoppedAgentSessionIds.delete(featureId);
-    return this.messageRepo.deleteByFeatureId(featureId);
+    return this.stateManager.clearMessages(
+      featureId,
+      this.sessions,
+      this.stoppedAgentSessionIds,
+      (sessionId) => this.stopSession(sessionId)
+    );
   }
 
   async getSession(sessionId: string): Promise<InteractiveSession | null> {
-    return this.sessionRepo.findById(sessionId);
+    return this.stateManager.getSession(sessionId);
   }
 
   subscribe(sessionId: string, onChunk: (chunk: StreamChunk) => void): UnsubscribeFn {
-    const state = this.sessions.get(sessionId);
-    if (!state) {
-      // eslint-disable-next-line @typescript-eslint/no-empty-function
-      return () => {};
-    }
-    state.subscribers.add(onChunk);
-    return () => state.subscribers.delete(onChunk);
+    return this.notifier.subscribeBySession(sessionId, this.sessions, onChunk);
   }
 
   // ---------------------------------------------------------------------------
@@ -859,395 +146,51 @@ export class InteractiveSessionService implements IInteractiveSessionService {
     model?: string,
     agentType?: string
   ): Promise<InteractiveMessage> {
-    // 1. Persist user message to DB immediately — this is the source of truth
-    const now = new Date();
-    const userMsg: InteractiveMessage = {
-      id: crypto.randomUUID(),
+    return this.turnExecutor.sendUserMessage(
       featureId,
-      role: InteractiveMessageRole.user,
       content,
-      createdAt: now,
-      updatedAt: now,
-    };
-    await this.messageRepo.create(userMsg);
-
-    // 2. Find active session for this feature
-    let state = this.findActiveStateForFeature(featureId);
-
-    // If the caller requested a different model/agent than the running session,
-    // silently stop the current session so a new one boots with the new config.
-    // Also clear the cached agentSessionId so we create a fresh SDK session
-    // instead of resuming the old one (which would keep the old model).
-    if (state && model && state.model !== model) {
-      await this.stopSession(state.sessionId);
-      this.stoppedAgentSessionIds.delete(featureId);
-      state = undefined;
-    } else if (state && agentType && state.agentType !== agentType) {
-      await this.stopSession(state.sessionId);
-      this.stoppedAgentSessionIds.delete(featureId);
-      state = undefined;
-    }
-
-    if (state) {
-      const dbSession = await this.sessionRepo.findById(state.sessionId);
-      if (dbSession?.status === InteractiveSessionStatus.ready) {
-        // Session ready — send to agent (guarded: one turn at a time)
-        this.resetTimer(state);
-        await this.sessionRepo.updateLastActivity(state.sessionId, now);
-        if (state.turnInProgress) {
-          state.turnQueue.push(content);
-        } else {
-          state.turnInProgress = true;
-          void this.executeAndPersistTurn(state, content);
-        }
-      } else if (dbSession?.status === InteractiveSessionStatus.booting) {
-        // Session booting — queue the message
-        state.pendingUserContent = content;
-      }
-    } else {
-      // No in-memory session — check DB for an orphaned active session (e.g. after
-      // service restart / hot-reload) and mark it stopped before booting a new one.
-      // The agentSessionId is persisted in DB so startSession will pick it up for
-      // SDK session resumption.
-      const dbSession = await this.sessionRepo.findByFeatureId(featureId);
-      if (
-        dbSession &&
-        (dbSession.status === InteractiveSessionStatus.ready ||
-          dbSession.status === InteractiveSessionStatus.booting)
-      ) {
-        await this.sessionRepo.updateStatus(
-          dbSession.id,
-          InteractiveSessionStatus.stopped,
-          new Date()
-        );
-      }
-
-      // Boot a new session — startSession will find the agentSessionId from DB
-      const session = await this.startSession(featureId, worktreePath, model, agentType);
-      const newState = this.sessions.get(session.id);
-      if (newState) {
-        newState.pendingUserContent = content;
-      }
-    }
-
-    return userMsg;
+      worktreePath,
+      this.sessions,
+      this.stoppedAgentSessionIds,
+      (fId, wPath, m, aType) => this.startSession(fId, wPath, m, aType),
+      (fId) => this.stateManager.findActiveStateForFeature(fId, this.sessions),
+      model,
+      agentType
+    );
   }
 
   async getChatState(featureId: string): Promise<ChatState> {
-    // DB messages
-    const messages = await this.messageRepo.findByFeatureId(featureId);
-
-    // Find active in-memory session
-    const state = this.findActiveStateForFeature(featureId);
-    let sessionStatus: string | null = null;
-    let streamingText: string | null = null;
-    let sessionInfo: ChatState['sessionInfo'] = null;
-
-    if (state) {
-      const dbSession = await this.sessionRepo.findById(state.sessionId);
-      sessionStatus = dbSession?.status ?? null;
-      if (state.currentAssistantBuffer) {
-        streamingText = state.currentAssistantBuffer;
-      }
-      // Resolve model display: explicit override > default
-      const displayModel = state.model ?? 'claude-sonnet-4-6';
-
-      const usage = await this.sessionRepo.getUsage(state.sessionId);
-      sessionInfo = {
-        pid: null, // SDK manages process internally
-        sessionId: state.agentSessionId ?? state.sessionId,
-        model: displayModel,
-        startedAt: dbSession?.startedAt
-          ? new Date(dbSession.startedAt as unknown as string).toISOString()
-          : new Date().toISOString(),
-        idleTimeoutMinutes: Math.round(this.getTimeoutMs() / 60_000),
-        lastActivityAt: dbSession?.lastActivityAt
-          ? new Date(dbSession.lastActivityAt as unknown as string).toISOString()
-          : new Date().toISOString(),
-        totalCostUsd: usage?.totalCostUsd ?? null,
-        totalInputTokens: usage?.totalInputTokens ?? null,
-        totalOutputTokens: usage?.totalOutputTokens ?? null,
-      };
-    } else {
-      // No in-memory state — check DB for last session (e.g. after server restart / hot-reload)
-      const latest = await this.sessionRepo.findByFeatureId(featureId);
-      if (latest) {
-        sessionStatus = latest.status as string;
-        // Show DB info even without live process (process was lost on restart)
-        if (
-          latest.status !== InteractiveSessionStatus.stopped &&
-          latest.status !== InteractiveSessionStatus.error
-        ) {
-          const latestUsage = await this.sessionRepo.getUsage(latest.id);
-          sessionInfo = {
-            pid: null,
-            sessionId: latest.id,
-            model: null,
-            startedAt: latest.startedAt
-              ? new Date(latest.startedAt as unknown as string).toISOString()
-              : new Date().toISOString(),
-            idleTimeoutMinutes: Math.round(this.getTimeoutMs() / 60_000),
-            lastActivityAt: latest.lastActivityAt
-              ? new Date(latest.lastActivityAt as unknown as string).toISOString()
-              : new Date().toISOString(),
-            totalCostUsd: latestUsage?.totalCostUsd ?? null,
-            totalInputTokens: latestUsage?.totalInputTokens ?? null,
-            totalOutputTokens: latestUsage?.totalOutputTokens ?? null,
-          };
-        }
-      }
-    }
-
-    // Resolve turn status from DB
-    let turnStatus = 'idle';
-    const activeState = state;
-    if (activeState) {
-      const statuses = await this.sessionRepo.getTurnStatuses([featureId]);
-      turnStatus = statuses.get(featureId) ?? 'idle';
-    } else {
-      // Check DB for the latest session's turn status
-      const latest = await this.sessionRepo.findByFeatureId(featureId);
-      if (latest) {
-        const statuses = await this.sessionRepo.getTurnStatuses([featureId]);
-        turnStatus = statuses.get(featureId) ?? 'idle';
-      }
-    }
-
-    // Include pending interaction if one exists
-    const pendingInteraction = state?.pendingInteraction ?? null;
-
-    return { messages, sessionStatus, streamingText, sessionInfo, turnStatus, pendingInteraction };
+    return this.chatStateBuilder.getChatState(featureId, this.sessions);
   }
 
   subscribeByFeature(featureId: string, onChunk: (chunk: StreamChunk) => void): UnsubscribeFn {
-    // Subscribe at the feature level so the callback survives session restarts.
-    // When a session dies (idle timeout, error) and a new one boots, the SSE
-    // connection keeps receiving events from the new session automatically.
-    let subs = this.featureSubscribers.get(featureId);
-    if (!subs) {
-      subs = new Set();
-      this.featureSubscribers.set(featureId, subs);
-    }
-    subs.add(onChunk);
-    return () => {
-      subs!.delete(onChunk);
-      if (subs!.size === 0) {
-        this.featureSubscribers.delete(featureId);
-      }
-    };
+    return this.notifier.subscribeByFeature(featureId, onChunk);
   }
 
   async stopByFeature(featureId: string): Promise<void> {
-    const state = this.findActiveStateForFeature(featureId);
-    if (!state) return;
-    await this.stopSession(state.sessionId);
+    return this.stateManager.stopByFeature(featureId, this.sessions, this.stoppedAgentSessionIds);
   }
 
   async markRead(featureId: string): Promise<void> {
-    // Find the active session for this feature and clear unread status
-    const state = this.findActiveStateForFeature(featureId);
-    if (state) {
-      void this.sessionRepo.updateTurnStatus(state.sessionId, 'idle');
-      return;
-    }
-    // Fallback: check DB for the latest active session
-    const latest = await this.sessionRepo.findByFeatureId(featureId);
-    if (latest) {
-      void this.sessionRepo.updateTurnStatus(latest.id, 'idle');
-    }
+    return this.stateManager.markRead(featureId, this.sessions);
   }
 
   async getTurnStatuses(featureIds: string[]): Promise<Map<string, string>> {
-    return this.sessionRepo.getTurnStatuses(featureIds);
+    return this.stateManager.getTurnStatuses(featureIds);
   }
 
   async getAllActiveTurnStatuses(): Promise<Map<string, string>> {
-    return this.sessionRepo.getAllActiveTurnStatuses();
+    return this.stateManager.getAllActiveTurnStatuses();
   }
 
   async respondToInteraction(featureId: string, answers: Record<string, string>): Promise<void> {
-    const state = this.findActiveStateForFeature(featureId);
-    if (!state?.pendingInteraction || !state.pendingInteractionResolver) {
-      throw new Error(`No pending interaction for feature ${featureId}`);
-    }
-
-    // Persist the user's answers as a structured user message.
-    // The {{interaction}} prefix lets the frontend detect and render it
-    // as a compact green bubble instead of a regular text message.
-    const interactionPayload = {
-      questions: state.pendingInteraction.questions.map((q) => ({
-        header: q.header,
-        question: q.question,
-      })),
-      answers,
-    };
-    const now = new Date();
-    const userMsg: InteractiveMessage = {
-      id: crypto.randomUUID(),
-      featureId: state.featureId,
-      sessionId: state.sessionId,
-      role: InteractiveMessageRole.user,
-      content: `{{interaction}}${JSON.stringify(interactionPayload)}`,
-      createdAt: now,
-      updatedAt: now,
-    };
-    await this.messageRepo.create(userMsg);
-
-    // Resolve the Promise that the canUseTool callback is awaiting.
-    // This unblocks the SDK stream — the agent resumes with the user's answers.
-    state.pendingInteractionResolver(answers);
-
-    // Clear pending interaction state
-    state.pendingInteraction = null;
-    state.pendingInteractionResolver = null;
-
-    // Update turn status back to processing
-    void this.sessionRepo.updateTurnStatus(state.sessionId, 'processing');
-
-    // Clear the "Waiting for your response..." log
-    state.subscribers.forEach((sub) => sub({ delta: '', done: false }));
-  }
-
-  /**
-   * Build the onUserQuestion callback for a session.
-   * Called by the SDK's canUseTool when the agent invokes AskUserQuestion.
-   * Returns a Promise that doesn't resolve until the user submits their answers.
-   */
-  private buildOnUserQuestionCallback(state: SessionState) {
-    return async (interaction: UserInteractionData): Promise<Record<string, string>> => {
-      // Flush any accumulated assistant text as a separate message BEFORE
-      // the interaction. This ensures the agent's question text appears
-      // above the green answer bubble in the conversation history.
-      if (state.currentAssistantBuffer.trim()) {
-        const now = new Date();
-        const msg: InteractiveMessage = {
-          id: crypto.randomUUID(),
-          featureId: state.featureId,
-          sessionId: state.sessionId,
-          role: InteractiveMessageRole.assistant,
-          content: state.currentAssistantBuffer,
-          createdAt: now,
-          updatedAt: now,
-        };
-        await this.messageRepo.create(msg);
-        state.currentAssistantBuffer = '';
-        state.toolEventsLog = [];
-
-        // Notify subscribers so the frontend picks up the new message
-        state.subscribers.forEach((sub) => sub({ delta: '', done: true }));
-        // Small delay so the refetch completes before the interaction appears
-        await new Promise<void>((r) => setTimeout(r, 100));
-      }
-
-      // Store the interaction data for the frontend
-      state.pendingInteraction = interaction;
-
-      // Update turn status so the dot indicator shows amber
-      void this.sessionRepo.updateTurnStatus(state.sessionId, 'awaiting_input');
-
-      // Notify subscribers so SSE pushes the interaction to the frontend
-      state.subscribers.forEach((sub) =>
-        sub({
-          delta: '',
-          done: false,
-          log: 'Waiting for your response...',
-          interaction,
-        })
-      );
-
-      // Create a Promise that will be resolved when the user calls respondToInteraction
-      return new Promise<Record<string, string>>((resolve) => {
-        state.pendingInteractionResolver = resolve;
-      });
-    };
-  }
-
-  /** Find the in-memory state for an active session for a feature. */
-  private findActiveStateForFeature(featureId: string): SessionState | undefined {
-    for (const state of this.sessions.values()) {
-      if (state.featureId === featureId) return state;
-    }
-    return undefined;
+    return this.turnExecutor.respondToInteraction(featureId, answers, (fId) =>
+      this.stateManager.findActiveStateForFeature(fId, this.sessions)
+    );
   }
 
   // ---------------------------------------------------------------------------
-  // Agent resolution helpers
-  // ---------------------------------------------------------------------------
-
-  /** Resolve the agent type from an explicit override or settings. */
-  private resolveAgentType(agentTypeOverride?: string): AgentType {
-    if (agentTypeOverride) {
-      return agentTypeOverride as AgentType;
-    }
-    if (hasSettings()) {
-      return getSettings().agent.type;
-    }
-    return AgentType.ClaudeCode;
-  }
-
-  /** Resolve the auth config from settings, with a safe fallback. */
-  private resolveAuthConfig(): AgentConfig {
-    if (hasSettings()) {
-      return getSettings().agent;
-    }
-    // Fallback for when settings haven't been initialized yet
-    return {
-      type: AgentType.ClaudeCode,
-      authMethod: AgentAuthMethod.Session,
-    };
-  }
-
-  // ---------------------------------------------------------------------------
-  // Tool detail extraction
-  // ---------------------------------------------------------------------------
-
-  /**
-   * Persist a tool/system event as its own assistant message in the DB.
-   * Each event gets its own bubble in the chat thread.
-   */
-  private async persistToolEvent(
-    state: SessionState,
-    label: string,
-    detail?: string
-  ): Promise<void> {
-    try {
-      const content = detail ? `**${label}** \`${detail}\`` : `**${label}**`;
-      const msg: InteractiveMessage = {
-        id: crypto.randomUUID(),
-        featureId: state.featureId,
-        sessionId: state.sessionId,
-        role: InteractiveMessageRole.assistant,
-        content,
-        createdAt: new Date(),
-        updatedAt: new Date(),
-      };
-      await this.messageRepo.create(msg);
-    } catch {
-      // Non-critical — don't fail the turn for a tool event
-    }
-  }
-
-  // ---------------------------------------------------------------------------
-  // Event dispatch
-  // ---------------------------------------------------------------------------
-
-  /**
-   * Dispatch a StreamChunk to all subscribers for a session.
-   *
-   * Sends to both session-level subscribers (legacy, for sessionId-based
-   * subscribe()) and feature-level subscribers (for SSE connections that
-   * must survive session restarts).
-   */
-  private notify(state: SessionState, chunk: StreamChunk): void {
-    state.subscribers.forEach((sub) => sub(chunk));
-    const featureSubs = this.featureSubscribers.get(state.featureId);
-    if (featureSubs) {
-      featureSubs.forEach((sub) => sub(chunk));
-    }
-  }
-
-  // ---------------------------------------------------------------------------
-  // Timer helpers
+  // Timer helpers (shared by sub-classes via bound callback)
   // ---------------------------------------------------------------------------
 
   /** Start or restart the idle timeout timer for a session. */
@@ -1273,12 +216,5 @@ export class InteractiveSessionService implements IInteractiveSessionService {
     const settings = getSettings();
     const minutes = settings.interactiveAgent?.autoTimeoutMinutes ?? 15;
     return minutes * 60 * 1000;
-  }
-
-  /** Read the concurrent session cap from settings or fall back to default. */
-  private getCap(): number {
-    if (!hasSettings()) return DEFAULT_CAP;
-    const settings = getSettings();
-    return settings.interactiveAgent?.maxConcurrentSessions ?? DEFAULT_CAP;
   }
 }
